+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 #Fixed
 - Dashboard: the number of papers assigned to the copy editor includes obsolete papers
+- RT#149643: the paper owner can see reviewers names.  
 
 ## 1.0.26 - 2022-03-01
 # Fixed